--- app/(marketing)/blog/[...slug]/head.tsx
@@ -1,12 +1,0 @@
-import MdxHead from "@/components/docs/mdx-head"
-
-export default function Head({ params }) {
-  return (
-    <MdxHead
-      params={params}
-      og={{
-        type: "Blog Post",
-      }}
-    />
-  )
-}
