+++ app/(marketing)/blog/[slug]/head.tsx
