+++ app/(marketing)/blog/[slug]/page.tsx
@@ -1,4 +1,4 @@
-import PostPage from '@/components/BlogPostPage'
+import BlogPostPage from '@/components/blog/BlogPostPage'
 import PreviewPostPage from '@/components/blog/PreviewPostPage'
 import { PreviewSuspense } from 'components/PreviewSuspense'
 import {
@@ -12,36 +12,39 @@ export async function generateStaticParams() {
   return await getAllPostsSlugs()
 }
 
-export default async function SlugRoute({
-  params,
-}: {
-  params: { slug: string }
-}) {
-  // Start fetching settings early, so it runs in parallel with the post query
+interface PostPageProps {
+  params: {
+    slug: string
+  }
+}
+
+export default async function PostPage({ params }: PostPageProps) {
   const settings = getSettings()
+   console.log(params.slug)
+
+
 
   if (previewData()) {
     const token = previewData().token || null
     const data = getPostAndMoreStories(params.slug, token)
+    console.log(data); 
     return (
-      <PreviewSuspense
-        fallback={
-          <PostPage
-            loading
-            preview
-            data={await data}
-            settings={await settings}
-          />
-        }
-      >
-        <PreviewPostPage token={token} slug={params.slug} />
-      </PreviewSuspense>
-    )
-  }
-
+    <PreviewSuspense
+      fallback={
+        <BlogPostPage
+          loading
+          preview
+          data={await data}
+          settings={await settings}
+        />
+      }
+    >
+      <PreviewPostPage token={token} slug={params.slug} />
+    </PreviewSuspense>
+  )
+} 
   const data = getPostAndMoreStories(params.slug)
-  return <PostPage data={await data} settings={await settings} />
+  return <BlogPostPage data={await data} settings={await settings} />
 }
-
 // FIXME: remove the `revalidate` export below once you've followed the instructions in `/pages/api/revalidate.ts`
 export const revalidate = 1
