+++ app/(marketing)/blog/page.tsx
@@ -1,71 +1,32 @@
-import Link from "next/link"
-import { compareDesc } from "date-fns"
-import { allPosts } from "contentlayer/generated"
 
-import { formatDate } from "@/lib/utils"
-import Image from "next/image"
+import IndexPage from '@/components/blog/IndexPage'
+import PreviewIndexPage from '@/components/blog/PreviewIndexPage'
+import { PreviewSuspense } from 'components/PreviewSuspense'
+import { compareDesc } from "date-fns"
+import { previewData } from 'next/headers'
+// import { allPosts } from "contentlayer/generated"
+import { getAllPosts, getSettings } from 'lib/sanity.client'
+
 
 export default async function BlogPage() {
-  const posts = allPosts
-    .filter((post) => post.published)
-    .sort((a, b) => {
-      return compareDesc(new Date(a.date), new Date(b.date))
-    })
+  // Fetch queries in parallel
+  const [settings, posts] = await Promise.all([getSettings(), getAllPosts()])
+  
+  if (previewData()) {
+    const token = previewData().token || null
 
-  return (
-    <div className="container max-w-4xl py-6 lg:py-10">
-      <div className="flex flex-col items-start gap-4 md:flex-row md:justify-between md:gap-8">
-        <div className="flex-1 space-y-4">
-          <h1 className="inline-block text-4xl font-extrabold tracking-tight text-slate-900 lg:text-5xl">
-            Blog
-          </h1>
-          <p className="text-xl text-slate-600">
-            A blog built using Contentlayer. Posts are written in MDX.
-          </p>
-        </div>
-        <Link
-          href="/guides"
-          className="relative inline-flex h-11 items-center rounded-md border border-slate-900 bg-white px-8 py-2 text-center font-medium text-slate-900 transition-colors hover:bg-slate-900 hover:text-slate-100 focus:outline-none focus:ring-2 focus:ring-brand-500 focus:ring-offset-2"
-        >
-          Build your own
-        </Link>
-      </div>
-      <hr className="my-8 border-slate-200" />
-      {posts?.length ? (
-        <div className="grid gap-10 sm:grid-cols-2">
-          {posts.map((post, index) => (
-            <article
-              key={post._id}
-              className="group relative flex flex-col space-y-2"
-            >
-              {post.image && (
-                <Image
-                  src={post.image}
-                  alt={post.title}
-                  width={804}
-                  height={452}
-                  className="rounded-md border border-slate-200 bg-slate-200 transition-colors group-hover:border-slate-900"
-                  priority={index <= 1}
-                />
-              )}
-              <h2 className="text-2xl font-extrabold">{post.title}</h2>
-              {post.description && (
-                <p className="text-slate-600">{post.description}</p>
-              )}
-              {post.date && (
-                <p className="text-sm text-slate-600">
-                  {formatDate(post.date)}
-                </p>
-              )}
-              <Link href={post.slug} className="absolute inset-0">
-                <span className="sr-only">View Article</span>
-              </Link>
-            </article>
-          ))}
-        </div>
-      ) : (
-        <p>No posts published.</p>
-      )}
-    </div>
-  )
+    return (
+      <PreviewSuspense
+        fallback={
+          <IndexPage loading preview posts={posts} settings={settings} />
+        }
+      >
+        <PreviewIndexPage token={token} />
+      </PreviewSuspense>
+    )
+  }
+
+  return <IndexPage posts={posts} settings={settings} />
 }
+
+export const revalidate = 1
