+++ app/layout.tsx
@@ -1,14 +1,40 @@
+import { Inter as FontSans } from "@next/font/google"
 import 'tailwindcss/tailwind.css'
+import "@/styles/globals.css"
 
-export default function RootLayout({
-  children,
-}: {
+
+import { cn } from "@/lib/utils"
+import { Toaster } from "@/ui/toast"
+import { Help } from "@/components/help"
+import { Analytics } from "@/components/analytics"
+import { TailwindIndicator } from "@/components/tailwind-indicator"
+
+const fontSans = FontSans({
+  subsets: ["latin"],
+  variable: "--font-inter",
+})
+
+interface RootLayoutProps {
   children: React.ReactNode
-}) {
+}
+
+export default function RootLayout({ children }: RootLayoutProps) {
   return (
-    <html lang="en">
+    <html
+      lang="en"
+      className={cn(
+        "bg-white font-sans text-slate-900 antialiased",
+        fontSans.variable
+      )}
+    >
       <head />
-      <body className="bg-white text-black">{children}</body>
+      <body className="min-h-screen">
+        {children}
+        <Analytics />
+        <Help />
+        <Toaster position="bottom-right" />
+        <TailwindIndicator />
+      </body>
     </html>
   )
 }
