--- components/BlogContainer.tsx
@@ -1,3 +1,0 @@
-export default function BlogContainer({ children }) {
-  return <div className="container mx-auto px-5">{children}</div>
-}
