+++ components/GeneralContainer.tsx
@@ -1,0 +1,3 @@
+export default function GeneralContainer({ children }) {
+  return <div className="container max-w-4xl py-6 lg:py-10">{children}</div>
+}
