+++ components/blog/HeroPost.tsx
@@ -1,6 +1,6 @@
 import AuthorAvatar from 'components/AuthorAvatar'
 import CoverImage from 'components/CoverImage'
-import Date from 'components/PostDate'
+import Date from '@/components/blog/PostDate'
 import type { Post } from 'lib/sanity.queries'
 import Link from 'next/link'
