+++ components/blog/PostHeader.tsx
@@ -1,7 +1,7 @@
 import Avatar from 'components/AuthorAvatar'
 import CoverImage from 'components/CoverImage'
-import Date from 'components/PostDate'
-import PostTitle from 'components/PostTitle'
+import PostTitle from '@/components/blog/PostTitle'
+import { formatDate } from "@/lib/utils"
 import type { Post } from 'lib/sanity.queries'
 
 export default function PostHeader(
@@ -10,6 +10,11 @@ export default function PostHeader(
   const { title, coverImage, date, author, slug } = props
   return (
     <>
+    {date && (
+      <time dateTime={date} className="block text-sm text-slate-600">
+        Published on {formatDate(date)}
+      </time>
+    )}
       <PostTitle>{title}</PostTitle>
       <div className="hidden md:mb-12 md:block">
         {author && <Avatar name={author.name} picture={author.picture} />}
@@ -20,9 +25,6 @@ export default function PostHeader(
       <div className="mx-auto max-w-2xl">
         <div className="mb-6 block md:hidden">
           {author && <Avatar name={author.name} picture={author.picture} />}
-        </div>
-        <div className="mb-6 text-lg">
-          <Date dateString={date} />
         </div>
       </div>
     </>
