+++ components/blog/PostTitle.tsx
@@ -1,0 +1,7 @@
+export default function PostTitle({ children }) {
+  return (
+    <h1 className="mt-2 inline-block text-4xl font-extrabold leading-tight text-slate-900 lg:text-5xl">
+      {children}
+    </h1>
+  )
+}
