+++ next.config.mjs
@@ -1,5 +1,5 @@
 /** @type {import('next').NextConfig} */
-import { withContentlayer } from "next-contentlayer"
+// import { withContentlayer } from "next-contentlayer"
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
@@ -28,4 +28,4 @@ const nextConfig = {
   },
 }
 
-export default withContentlayer(nextConfig)
+export default nextConfig
